+++ solidity/contracts/token/SmartTokenController.sol
@@ -45,7 +45,7 @@ contract SmartTokenController is TokenHolder {
 
     /**
         @dev allows transferring the token ownership
-        the new owner still need to accept the transfer
+        the new owner needs to accept the transfer
         can only be called by the contract owner
 
         @param _newOwner    new token owner