+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Converters:
 * Removed the setRegistry function
 * getReturn now returns the fee as a separate return value
 * Converter owner can no longer withdraw connector tokens while the converter is active
+* Converter owner can no longer transfer the token ownership
 * Added a dedicated `upgrade` function for easier upgrades
 
 Registry: